Working version of the index composition example

+++ src/main/java/org/aa/olympus/api/UpdateContext.java
@@ -6,5 +6,5 @@ public interface UpdateContext {
 
   LocalDateTime getTime();
 
-  long getUpdateId();
+  int getUpdateId();
 }
